# swagger.yaml
openapi: "3.0.2"
info:
  title: "Puppy Shop"
  version: "1.0"
servers:
  - url: http://localhost:10778/
    description: shop server
paths:
...
  /carts:
    get:
      description: get open cart for user
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/cart"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    post:
      description: add new cart for user
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/cart"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /carts/{id}:
    get:
      description: get total cart price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 4
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  totalCartPrice:
                    type: integer
                    example: 345
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    put:
      description: change status of cart for user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 4
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/cart"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /carts/{cart_id}/items: 
    get:
      description: get cart items for user
      security:
        - token: []
          refreshToken: []
      parameters:
        - in: path
          name: cart_id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          $ref: "#/components/responses/cartItems"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    post:
      description: add item to cart
      security:
        - token: []
          refreshToken: []
      parameters:
        - in: path
          name: cart_id
          required: true
          schema:
            type: integer
            example: 1 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  example: 4
                product_amount:
                  type: integer
                  example: 2
                product_total_price:
                  type: integer
                  example: 200
      responses:
        "200":
          $ref: "#/components/responses/cartItemsUpdate"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    delete:
      description: delete all items in cart
      parameters:
        - name: cart_id
          in: path
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 4
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/cartItems"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /carts/{cart_id}/items/{item_id}: 
    delete:
      description: delete item from cart
      security:
        - token: []
          refreshToken: []
      parameters:
        - in: path
          name: cart_id
          required: true
          schema:
            type: integer
            example: 1
        - in: path
          name: item_id
          description: cart item id
          required: true
          schema:
            type: integer
            example: 2
      responses:
        "200":
          $ref: "#/components/responses/cartItems"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    put:
      description: edit quantity of an item
      security:
        - token: []
          refreshToken: []
      parameters:
        - name: cart_id
          in: path
          required: true
          schema:
            type: integer
            example: 1 
        - name: item_id
          in: path
          required: true
          schema:
            type: integer
            example: 4     
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_amount:
                  type: integer
                  example: 2
                product_total_price:
                  type: integer
                  example: 200
      responses:
        "200":
          $ref: "#/components/responses/cartItemsUpdate"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
...
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
    refreshToken:
      type: apiKey
      in: header
      name: refreshToken
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: true
              msg:
                type: string
                example: error
    unauthorized:
      description: unauthorized action
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: true
              msg:
                type: string
                example: unauthorized action
...
    cart:
      description: open/new cart
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              cart:
                type: object
                properties:
                  id: 
                    type: integer
                    example: 1 
                  user_id: 
                    type: integer
                    example: 2 
                  create_date: 
                    type: string
                    example: "2021-05-05"
                  status: 
                    type: boolean
                    example: true
    cartItems:
      description: cart items
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              cartItems:
                type: array
                items:
                  type: object
                  properties:
                    cartItem_id: 
                      type: integer
                      example: 1
                    name: 
                      type: string
                      example: item1
                    price:
                      type: integer
                      example: 35
                    product_id:
                      type: integer
                      example: 3
                    product_amount:
                      type: integer
                      example: 2
                    product_total_price:
                      type: integer
                      example: 70
                    cart_id: 
                      type: integer
                      example: 4
                    image: 
                      type: string
                      example: url.string1 
    cartItemsUpdate:
      description: cart items add/edit items
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              cartItems:
                type: array
                items:
                  type: object
                  properties:
                    cartItem_id: 
                      type: integer
                      example: 1
                    name: 
                      type: string
                      example: item1
                    price:
                      type: integer
                      example: 35
                    product_id:
                      type: integer
                      example: 3
                    product_amount:
                      type: integer
                      example: 2
                    product_total_price:
                      type: integer
                      example: 70
                    cart_id: 
                      type: integer
                      example: 4
                    image: 
                      type: string
                      example: url.string1 
              totalCartPrice:
                type: integer
                example: 345
...